Update facebook user mapping to include both email and name, and fix issues where they were confused.
This will require anywone with an existing login to remove B-Side from app approval on Facebook: go to user setgings and select "Apps" from the left nav. Serach for B-Sides in "Apps you use" may require expanding the list), and click the "x" at the end of the row to remove any existing authorization.